Move go_rest_server generator and outputs to sensible locations
