+++ server/genrest/README.md
@@ -1,0 +1,4 @@
+# server/genrest
+
+This directory contains auto-generated files used to implement a REST endpoint
+for Showcase services.
