+++ util/cmd/protoc-gen-go_rest_server/README.md
@@ -1,0 +1,5 @@
+# protoc-gen-go_rest_server
+
+This directory contains the `protoc` plugin `go_rest_server` used to generate a Go
+HTTP server that can translate REST requests and responses to and from protocol
+buffer messages used by the Showcase gRPC server.
