+++ util/genrest/README.md
@@ -1,0 +1,3 @@
+# util/genrest
+
+This directory contains the core functionality for the `protoc` plugin `go_rest_server`, which implements a REST endpoint for Showcase services.
