+++ util/genrest/genserver.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package genrest
 
 import (
 	"fmt"
@@ -24,6 +24,7 @@ import (
 )
 
 func Gen(genReq *plugin.CodeGeneratorRequest) (*plugin.CodeGeneratorResponse, error) {
+	// TODO(vchudnov-g) Fill in this stub
 	fileName := "showcase-rest-sample-response.txt"
 
 	resp := &plugin.CodeGeneratorResponse{}
